Rename "ProgDialog.py" to "ShellDialog.py"
